requirements.txt
PyQt6>=6.4.0
python-vlc>=3.0.16120
sounddevice>=0.4.6
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
